+++ src/api/reserve.ts
@@ -13,23 +13,6 @@ interface ScreeningData {
   };
 }
 
-interface ReservationData {
-  reservationNumber: string;
-  title: string;
-  startDate: string;
-  timeRange: string;
-  auditoriumName: string;
-  ticketDetails: string;
-  totalPrice: string;
-  seats: Seat[];
-}
-
-interface Seat {
-  row: number;
-  number: number;
-  seatId: number;
-}
-
 async function getScreeningData(screeningId: number) {
   const response = await getAxios().get<ScreeningData>(`seats/${screeningId}`);
   console.log(response.data.message);
@@ -51,33 +34,4 @@ export const reserveLoader =
     await client.ensureQueryData(getScreeningDataQuery(+params.screeningId));
 
     return { screeningId: +params.screeningId };
-  };
-
-//-------------BOOKING-INFORMATION---------------
-async function getBookingData(bookingNumber: string) {
-  const response = await getAxios().get<ReservationData>(
-    `/reservation/${bookingNumber}`
-  );
-
-  return response.data;
-}
-
-export const getBookingDataQuery = (bookingNumber: string) =>
-  queryOptions({
-    queryKey: ['booking', bookingNumber],
-    queryFn: async () => await getBookingData(bookingNumber),
-  });
-
-export const bookingLoader =
-  (client: QueryClient) =>
-  async ({ params }: LoaderFunctionArgs) => {
-    const bookingNumber = params.reservationNum;
-
-    if (!bookingNumber) {
-      throw new Error('Bokingsnummer saknas');
-    }
-
-    await client.ensureQueryData(getBookingDataQuery(bookingNumber));
-
-    return { bookingNumber };
   };
